refact: create usdt and usdc config utils to be used in testnet and mainnet
