+++ packages/stg-evm-v2/devtools/config/mainnet/01/usdt-token.config.ts
@@ -1,51 +1,9 @@
-import { TokenName } from '@stargatefinance/stg-definitions-v2'
-
-import { OmniGraphHardhat, createContractFactory, createGetHreByEid } from '@layerzerolabs/devtools-evm-hardhat'
+import { OmniGraphHardhat } from '@layerzerolabs/devtools-evm-hardhat'
+import { Stage } from '@layerzerolabs/lz-definitions'
 import { OwnableNodeConfig } from '@layerzerolabs/ua-devtools'
 
-import { createGetAssetAddresses, getAssetNetworkConfig } from '../../../../ts-src/utils/util'
-import { getContractWithEid } from '../../utils'
-import { getChainsThatSupportsUsdtOftByDeployment } from '../../utils/utils.config'
-import { setMainnetStage } from '../utils'
-
-const fiatContract = { contractName: 'TetherTokenV2' }
-
-const tokenName = TokenName.USDT
+import buildUsdtTokenGraph from '../../utils/usdt-token.config.utils'
 
 export default async (): Promise<OmniGraphHardhat<OwnableNodeConfig, unknown>> => {
-    // Set the stage to mainnet
-    setMainnetStage()
-
-    // First let's create the HardhatRuntimeEnvironment objects for all networks
-    const getEnvironment = createGetHreByEid()
-    const contractFactory = createContractFactory(getEnvironment)
-    const getAssetAddresses = createGetAssetAddresses(getEnvironment)
-
-    // get the list of chains that supports usdt external deployments (oft usdt with external deployment)
-    const chains = getChainsThatSupportsUsdtOftByDeployment(true)
-    const contracts = await Promise.all(
-        chains.map(async (chain) => {
-            const usdtProxyAddress = await contractFactory(
-                getContractWithEid(chain.eid, {
-                    contractName: 'TransparentUpgradeableProxy',
-                    address: getAssetNetworkConfig(chain.eid, tokenName).address, // usdt asset address
-                })
-            )
-            const assetAddresses = await getAssetAddresses(chain.eid, [tokenName])
-            return {
-                contract: getContractWithEid(chain.eid, {
-                    ...fiatContract,
-                    address: usdtProxyAddress.contract.address,
-                }),
-                config: {
-                    owner: assetAddresses[tokenName],
-                },
-            }
-        })
-    )
-
-    return {
-        contracts,
-        connections: [],
-    }
+    return buildUsdtTokenGraph(Stage.MAINNET)
 }
